+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - :sparkles: Feat(math): add more block delimiter support for math
 - :sparkles: Feat(highlight): add full support for Chroma highlighting languages in the [list](https://gohugo.io/content-management/syntax-highlighting/#list-of-chroma-highlighting-languages)
 - :sparkles: Feat: add optional parameter `noreferrer` for the shortcode `link`
+- :sparkles: Feat: update echarts theme config
 - :wheelchair: Feat(accessibility): use `aria-hidden=true` on icons that AT should ignore
 - :bug: Fix: add function doc2unix to unify new lines symbol between Windows and Unix/Mac OS
 - :bug: Fix: author display error in post and markdown file